requirements.txt
streamlit
tensorflow
pillow
numpy>=1.23.5,<2.0.0
opencv-python-headless==4.8.1.78
pandas
groq
python-dotenv

